pyproject.toml
[tool.poetry]
name = "asset-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"aws-cdk.core" = "1.73.0"
"aws-cdk.pipelines" = "1.73.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
